envoi calcul moyenne au front

+++ src/main/java/com/projet/ecommerce/persistance/repository/AvisClientRepositoryCustom.java
@@ -2,5 +2,5 @@ package com.projet.ecommerce.persistance.repository;
 
 public interface AvisClientRepositoryCustom {
 
-	double averageByReferenceProduit(String reference);
+	float averageByReferenceProduit(String reference);
 }
